# Lab2/pzpi-22-3-shpak-denys-lab2/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.AuthCredentials:
    properties:
      password:
        description: |-
          Password is the password of the user
          Example: password123
        example: password123
        type: string
      username:
        description: |-
          Username is the name of the user
          Example: john_doe
        example: john_doe
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  termsOfService: http://swagger.io/terms/
  title: API Specification
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthCredentials'
      produces:
      - application/json
      responses: {}
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logs out a user by invalidating their token
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the provided details
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthCredentials'
      produces:
      - application/json
      responses: {}
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'Provide the Bearer token in the format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"